Fixed the issue of of music bars not animating on item selection and improvements for the now playing music bars.
Improvements for the Player transition animation.

Refactor and restructure.

+++ MusicApp/Models/Media/ItunesAPIResponse.swift
@@ -37,7 +37,7 @@ struct MediaResponse: Codable {
   let discNumber: Int?
   let trackCount: Int?
   let trackNumber: Int?
-  let trackTimeMillis: Int?
+  let trackTimeMillis: Double?
   let country: String?
   let currency: String?
   let primaryGenreName: String?
